Various fixes and type export additions (blockchain, VM)
