+++ packages/blockchain/src/blockchain.ts
@@ -17,8 +17,7 @@ import { genesisStateRoot } from './genesisStates'
 import {} from './utils'
 
 import type { Consensus } from './consensus'
-import type { GenesisState } from './genesisStates'
-import type { BlockchainInterface, BlockchainOptions, OnBlock } from './types'
+import type { BlockchainInterface, BlockchainOptions, GenesisState, OnBlock } from './types'
 import type { BlockData } from '@ethereumjs/block'
 import type { CliqueConfig } from '@ethereumjs/common'
 import type { BigIntLike, DB, DBObject } from '@ethereumjs/util'
