+++ packages/blockchain/src/index.ts
@@ -1,4 +1,5 @@
 export { Blockchain } from './blockchain'
 export { CasperConsensus, CliqueConsensus, Consensus, EthashConsensus } from './consensus'
-export { BlockchainInterface, BlockchainOptions } from './types'
+export { DBOp, DBSaveLookups, DBSetBlockOrHeader, DBSetHashToNumber, DBSetTD } from './db/helpers'
+export * from './types'
 export * from './utils'
