+++ packages/devp2p/package.json
@@ -29,8 +29,14 @@
     "Martin Becze <mb@ethdev.com>",
     "Holger Drewes <holger.drewes@gmail.com>"
   ],
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "main": "dist/cjs/index.js",
+  "module": "dist/esm/index.js",
+  "exports": {
+    ".": {
+      "import": "./dist/esm/index.js",
+      "require": "./dist/cjs/index.js"
+    }
+  },
   "files": [
     "dist",
     "src"
