+++ packages/vm/src/index.ts
@@ -1,4 +1,5 @@
 export { Bloom } from './bloom/index'
 export { BlockBuilder, BuildStatus } from './buildBlock'
+export { encodeReceipt } from './runBlock'
 export * from './types'
 export { VM } from './vm'
